f-analytics - Added a watch on route change to then re-push the platformData + plus split out tests to reduce mock leakage.
