+++ packages/services/f-analytics/src/mixins/_tests/analytics.created.test.js
@@ -1,5 +1,6 @@
 import Vuex from 'vuex';
 import { shallowMount, createLocalVue } from '@vue/test-utils';
+import VueRouter from 'vue-router';
 import analyticsMixin from '../analytics.mixin.vue';
 import {
     defaultState,
@@ -9,56 +10,11 @@ import {
 
 const localVue = createLocalVue();
 localVue.use(Vuex);
+localVue.use(VueRouter);
 
 describe('Analytics', () => {
     afterEach(() => {
         jest.clearAllMocks();
-    });
-
-    describe('mounted ::', () => {
-        let preparePageMock;
-        let prepareAnalyticsMock;
-        let pushAnalyticsMock;
-
-        beforeEach(() => {
-            // Arrange
-            const component = {
-                render () {},
-                mixins: [analyticsMixin],
-                store: createStore()
-            };
-
-            // Mocks
-            preparePageMock = jest.spyOn(component.mixins[0].methods, 'preparePage').mockImplementationOnce(() => true);
-            prepareAnalyticsMock = jest.spyOn(component.mixins[0].methods, 'prepareAnalytics').mockImplementationOnce(() => true);
-            pushAnalyticsMock = jest.spyOn(component.mixins[0].methods, 'pushAnalytics').mockImplementationOnce(() => true);
-
-            // Act
-            shallowMount(
-                component,
-                {
-                    localVue,
-                    mocks: {
-                        $cookies
-                    }
-                }
-            );
-        });
-
-        it('should make a call to `preparePage`', () => {
-            // Assert
-            expect(preparePageMock).toHaveBeenCalled();
-        });
-
-        it('should make a call to `prepareAnalytics`', () => {
-            // Assert
-            expect(prepareAnalyticsMock).toHaveBeenCalled();
-        });
-
-        it('should make a call to `pushAnalytics`', () => {
-            // Assert
-            expect(pushAnalyticsMock).toHaveBeenCalled();
-        });
     });
 
     describe('created ::', () => {
@@ -73,14 +29,8 @@ describe('Analytics', () => {
                 mixins: [analyticsMixin],
                 store: createStore()
             };
-
-            // Mocks
-            jest.spyOn(component.mixins[0].methods, 'preparePage').mockImplementationOnce(() => true);
-            jest.spyOn(component.mixins[0].methods, 'prepareAnalytics').mockImplementationOnce(() => true);
-            jest.spyOn(component.mixins[0].methods, 'pushAnalytics').mockImplementationOnce(() => true);
+            component.mixins[0].mounted = jest.fn(() => true);
             storeUpdatePlatformDataSpy = jest.spyOn(component.mixins[0].methods, 'updatePlatformData').mockImplementationOnce(() => true);
-
-            // Spies
             registerStoreModuleSpy = jest.spyOn(component.mixins[0].methods, 'registerStoreModule');
             prepareServersideAnalyticsSpy = jest.spyOn(component.mixins[0].methods, 'prepareServersideAnalytics');
         });
