+++ packages/services/f-analytics/src/mixins/analytics.mixin.vue
@@ -18,6 +18,13 @@ export default {
         }
     },
 
+    watch: {
+        $route () {
+            this.prepareAnalytics();
+            this.pushAnalytics();
+        }
+    },
+
     created () {
         this.registerStoreModule();
         this.prepareServersideAnalytics();
